+++ include/rain/networking/request-response/worker.hpp
@@ -116,8 +116,6 @@ namespace Rain::Networking::RequestResponse {
 		// TODO: Better access restrictions.
 		protected:
 		// Only allow a subset of streams.
-		using SuperInterface::operator<<;
-		using SuperInterface::operator>>;
 		using SuperInterface::send;
 		using SuperInterface::recv;
 		using SuperInterface::recvResponse;
@@ -127,6 +125,9 @@ namespace Rain::Networking::RequestResponse {
 		virtual void streamInImpl(Tag<Interface>, Request &) {}
 
 		public:
+		using SuperInterface::operator<<;
+		using SuperInterface::operator>>;
+
 		// Establish pre/post-processing chains via Tags on stream operators.
 		virtual Interface &operator<<(Response &res) final override {
 			this->streamOutImpl(Tag<Interface>(), res);